AI/requirements.txt
# Flask Web Framework
Flask>=3.0.0
flask-cors>=4.0.0

# Environment variables
python-dotenv>=1.0.0

# Hugging Face & LLM
huggingface-hub>=0.20.0

# LangChain
langchain>=0.1.0
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0

# Vector Database
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# HTTP & API
requests>=2.31.0

# Text analysis
textstat>=0.7.0

# System utilities
psutil>=5.9.0

# NLP (optional but recommended)
spacy>=3.7.0
